remove MegatronTrainer
